feat(seo): split deals/[id] page for server metadata generation, add breadcrumb schema
