+++ app/marketplace/page.tsx
@@ -1,5 +1,16 @@
 import { redirect } from 'next/navigation'
 
+export async function generateMetadata() {
+  return {
+    title: 'Marketplace | Mercato',
+    description: 'Redirecting to Mercato deals browser...',
+    alternates: {
+      canonical: '/marketplace',
+    },
+    robots: { index: false, follow: true },
+  }
+}
+
 /** @deprecated Use `/deals` — kept for bookmarks and external links. */
 export default async function MarketplaceRedirectPage({
   searchParams,
